+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+- upgrade: update spring-boot from v2.7.5 to v3.0.0
+
 ## [0.6.0] - 2022-11-23
 
 ### Breaking