+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-12-05
+
 ### Breaking
 - upgrade: update spring-boot from v2.7.5 to v3.0.0
 - feat: immutable configuration properties
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - upgrade: update checkstyle from v8.44 to v10.3.4
+
 
 ## [0.6.0] - 2022-11-23
 
@@ -30,12 +33,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - upgrade: update nebula project plugin from v9.6.3 to v10.0.1
 - upgrade: update spring-boot from v2.7.3 to v2.7.5
 
+
 ## [0.5.0] - 2022-10-21
 
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/theborakompanioni/tor-spring-boot-starter/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/theborakompanioni/tor-spring-boot-starter/compare/0.7.0...HEAD
+[0.7.0]: https://github.com/theborakompanioni/tor-spring-boot-starter/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/theborakompanioni/tor-spring-boot-starter/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/theborakompanioni/tor-spring-boot-starter/releases/tag/0.5.0
-